pyproject.toml
[project]
name = "gold_trading_one_trade_per_day"
version = "0.1.0"
description = "Event-driven GLD multi-agent scalping system using CrewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[litellm,tools,google-genai]==1.9.3",
    "pandas",
    "alpaca-py",
    "pytz",
]

[project.scripts]
gold_trading_one_trade_per_day = "gold_trading_one_trade_per_day.main:run"
run_crew = "gold_trading_one_trade_per_day.main:run"
run_shadow = "gold_trading_one_trade_per_day.main:run_shadow"
run_shadow_loop = "gold_trading_one_trade_per_day.main:run_shadow_loop"
run_paper = "gold_trading_one_trade_per_day.main:run_paper"
run_paper_smoke = "gold_trading_one_trade_per_day.main:run_paper_smoke"
run_live = "gold_trading_one_trade_per_day.main:run_live"
warmup = "gold_trading_one_trade_per_day.main:warmup"
benchmark_models = "gold_trading_one_trade_per_day.main:benchmark_models"
resume = "gold_trading_one_trade_per_day.main:resume"
reconcile = "gold_trading_one_trade_per_day.main:reconcile"
run_with_trigger = "gold_trading_one_trade_per_day.main:run_with_trigger"
train = "gold_trading_one_trade_per_day.main:train"
replay = "gold_trading_one_trade_per_day.main:replay"
test = "gold_trading_one_trade_per_day.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
